+++ src/cores/base_data_repository.py
@@ -1,12 +1,10 @@
 from abc import ABC, abstractmethod
+from typing import List
 from pathlib import Path
 import pandas as pd
-from interfaces.data_repository import DataRepositoryInterface
+from interfaces.data_repository_interface import DataRepositoryInterface
 
 class BaseDataRepository(DataRepositoryInterface):
-    def __init__(self, root_directory: Path) -> None:
-        self._root_directory = root_directory
-    
     def _ensure_path_exists(self, path: Path) -> None:
         """
         pathに指定したディレクトリを作成する
@@ -16,6 +14,6 @@ class BaseDataRepository(DataRepositoryInterface):
     def save(self, path: Path, df: pd.DataFrame) -> None:
         raise NotImplementedError
     @abstractmethod
-    def load(self, path: Path) -> pd.DataFrame:
+    def load(self, path: Path) -> object:
         raise NotImplementedError