+++ src/cores/csv_data_repository.py
@@ -1,16 +1,16 @@
 from pathlib import Path
+from typing import List
 import pandas as pd
 from cores.base_data_repository import BaseDataRepository
 
 
 class CSVDataRepository(BaseDataRepository):
-    def __init__(self, root_directory: Path) -> None:
-        self._root_directory = root_directory
 
     def save(self, path: Path, df: pd.DataFrame) -> None:
         """
         dfを保存する
         """
+        self._ensure_path_exists(path)
         df.to_csv(path, index=False)
 
     def load(self, path: Path) -> pd.DataFrame: